requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.11
jira==3.5.2
pywin32==306; sys_platform == 'win32'
msal==1.26.0
PyGithub==2.1.1
pandas==2.1.3
plotly==5.18.0
wakatime==1.0.0
pytz==2023.3